# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/category": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get paginated list of categories",
                "operationId": "d2f23b45653dc0dff84f355889899c25",
                "parameters": [
                    {
                        "name": "show",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a category by ID",
                "operationId": "c537f531f1c82505c335ed294c0b7ba7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/page": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get paginated list of pages",
                "operationId": "e0abae1b82c50f50e573e4bc84a55495",
                "parameters": [
                    {
                        "name": "show",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/page/{id}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get a page by ID",
                "operationId": "86dd6927fbebad7b064f7669f0602925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
        "/api/post": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get paginated list of posts",
                "operationId": "a8f335e0ced7872d7012086d0ecce9dd",
                "parameters": [
                    {
                        "name": "show",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/post/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a post by ID",
                "operationId": "edc5f5b547c313bb9fd4799b234b4c31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Pages",
            "description": "Pages"
        },
        {
            "name": "Posts",
            "description": "Posts"
        }
    ]
}